A lot of work around the discounting logic (fixes, refactoring and unit tests).
Added a new TriggerRequirements class to track what it takes to trigger a promotion (what items to buy, how many to buy, base on weight or quantities, apply case-rollup logic (Frank Liquor)).

It nests the trigger group requirements (e.g. you have to buy 100 cases of diet soda of which 10 cases must be diet coke).

Given the triggerQtys object (which contains a dictionary of sales by itemNid) it will return 'true' if the quantities are sufficient to trigger the requirements.

Added support for weekday promotions (if you take delivery on Monday, you'll get a discount - used to balance deliveries during the week)

Added a convenience property for an item called 'numberOfPrimaryPacks' - if you buy (1) of this alt-pack then that represents (e.g.) 1/4 of a primary pack