Added DepositService to compute line-item deposits (bottle and can deposits, CRV, carrier deposits, keg deposits, credits for pickup of empties, ...)
Refactored order lines to accept addition of a charge or credit (and enums to distinguish between the types of charges or credits) rather than having specific properties for each type of charge or credit.

Some bug fixes in the Money math (and added a divided:by method to control rounding of the result.